refactor: move database initialization and ISqlConnectionFactory to Infrastructure layer
